requirements.txt
transformers
torch
opencv-python
numpy
matplotlib
pillow
